Fixed bug in get_large_enough_prime
If the secret is a prime in STANDARD_PRIMES get_large_enough_prime returned the secret. Since the secret modulo itself is 0 this resulted in loss of the secret. This was fixed by making sure the prime needs to be GREATER than the secret and never equal to the secret.